+++ birdbox/microsite/models.py
@@ -140,6 +140,16 @@ class BaseProtocolPage(MetadataPageMixin, CacheAwareAbstractBasePage):
         ),
     )
 
+    canonical_rel = URLField(
+        blank=True,
+        help_text=mark_safe(
+            'Value to use within a &lt;link rel="canonical" ...&gt; tag in the head of this page, '
+            "to indicate a preferred URL for this page elsewhere. "
+            "Only set this if you know what you are doing. "
+            'See <a target="_blank" href="https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/rel#canonical">MDN docs</a>.'
+        ),
+    )
+
     settings_panels = Page.settings_panels + [
         FieldPanel("page_layout"),
         MultiFieldPanel(
@@ -153,9 +163,12 @@ class BaseProtocolPage(MetadataPageMixin, CacheAwareAbstractBasePage):
         ),
     ]
 
-    # Crudely drop the show_in_menus section. TODO: make this more elegant and less brittle
+    # Crudely drop the show_in_menus section, which we've moved into the settings_panels
+    # TODO: make this more elegant and less brittle
     promote_panels = Page.promote_panels[:-1]
     promote_panels += [FieldPanel("search_image")]
+
+    promote_panels += [FieldPanel("canonical_rel")]
 
     def has_menu_icon(self):
         return bool(self.menu_icon)